requirements.txt
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
Pygments==2.8.1
python-dotenv==0.15.0
rich==9.13.0
torch==1.8.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
